Refactor search for readability
Divide into smaller modules and add code comments
