+++ src/lib/scripts/search/adapters/proskomma-verse-provider.ts
@@ -1,26 +1,10 @@
 import type { SABProskomma } from '$lib/sab-proskomma';
-import {
-    SearchInterface,
-    SearchQueryBase,
-    type SearchCandidate,
-    type SearchOptions
-} from './application';
-import { BufferedReader } from './utils/buffered-reader';
-import { RegexHelpers, makeRegexPattern } from './utils/regex-helpers';
-import config from '$lib/data/config';
-import { extendStringProperty } from './utils/object-helpers';
+import { SearchInterface, type SearchCandidate } from '../application';
+import { BufferedReader } from '../utils/buffered-reader';
+import { RegexHelpers, makeRegexPattern } from '../utils/regex-helpers';
 
 export interface GQLBooks {
-    data: {
-        docSet: {
-            documents: {
-                id: string;
-                idParts: {
-                    type: string;
-                };
-            }[];
-        };
-    };
+    data: { docSet: { documents: { id: string; idParts: { type: string } }[] } };
 }
 
 export interface GQLBlockToken {
@@ -30,14 +14,7 @@ export interface GQLBlockToken {
 
 export interface GQLBlocks {
     data: {
-        document: {
-            bookCode: string;
-            mainSequence: {
-                blocks: {
-                    tokens: GQLBlockToken[];
-                }[];
-            };
-        };
+        document: { bookCode: string; mainSequence: { blocks: { tokens: GQLBlockToken[] }[] } };
     };
 }
 
@@ -56,7 +33,6 @@ function keywordToRegex(
     }
     return pattern;
 }
-
 function searchParams(
     keywords: string[],
     wholeWords: boolean = false,
@@ -66,11 +42,9 @@ function searchParams(
     const terms = keywords.map((w) => keywordToRegex(w, wholeWords, ignore, substitute));
     return `withMatchingChars: ["${terms.join('", "')}"]`;
 }
-
-/**
- * Get a chapter/verse reference string from Proskomma scopes
- */
-function chapterVerseFromScopes(scopes: string[]): string {
+/** * Get a chapter/verse reference string from Proskomma scopes */ function chapterVerseFromScopes(
+    scopes: string[]
+): string {
     let verse: string;
     let chapter: string;
     for (const s of scopes) {
@@ -86,11 +60,7 @@ function chapterVerseFromScopes(scopes: string[]): string {
     }
     return ref;
 }
-
-export const gqlSearchHelpers = {
-    keywordToRegex,
-    chapterVerseFromScopes
-};
+export const gqlSearchHelpers = { keywordToRegex, chapterVerseFromScopes };
 
 class ProskommaVerseProvider extends SearchInterface.VerseProvider {
     constructor(args: {
@@ -106,33 +76,26 @@ class ProskommaVerseProvider extends SearchInterface.VerseProvider {
         this.pk = args.pk;
         this.docSet = args.docSet;
         this.collection = args.collection;
-
         const tokens = RegexHelpers.wordsOf(args.searchPhrase);
         this.searchIsBlank = tokens.length === 0;
-
         this.searchParams = searchParams(tokens, args.wholeWords, args.ignore, args.substitute);
     }
-
     pk: SABProskomma;
     books: string[];
     nextBook: number;
     searchParams: string;
     searchIsBlank: boolean;
-
     docSet: string;
     collection: string;
-
     private verseReader = new BufferedReader({
         read: () => this.queryNextBook(),
         done: () => this.nextBook >= this.books.length
     });
-
     async getVerses(limit: number = 0): Promise<SearchCandidate[]> {
         if (this.searchIsBlank) return [];
         await this.ensureBooksSet();
         return await this.verseReader.read(limit);
     }
-
     async queryNextBook(): Promise<SearchCandidate[]> {
         await this.ensureBooksSet();
         if (this.nextBook >= this.books.length) {
@@ -142,13 +105,11 @@ class ProskommaVerseProvider extends SearchInterface.VerseProvider {
         this.nextBook++;
         return this.versesOfBook(book);
     }
-
     private async ensureBooksSet() {
         if (this.books == undefined) {
             await this.setBooks();
         }
     }
-
     async setBooks() {
         this.nextBook = 0;
         const queryData = await this.queryForBooks();
@@ -156,7 +117,6 @@ class ProskommaVerseProvider extends SearchInterface.VerseProvider {
             .filter((bk) => bk.idParts.type === 'book')
             .map((bk) => bk.id);
     }
-
     async versesOfBook(bookId: string): Promise<SearchCandidate[]> {
         const queryData = await this.queryForBlocks(bookId);
         const bookCode = queryData.data.document.bookCode;
@@ -165,7 +125,6 @@ class ProskommaVerseProvider extends SearchInterface.VerseProvider {
             .flat();
         return this.versesFromTokens(tokens, bookCode);
     }
-
     versesFromTokens(tokens: GQLBlockToken[], bookCode: string): SearchCandidate[] {
         const verseTexts = this.collectVerseTexts(tokens);
         return Object.keys(verseTexts).map((cv) => {
@@ -182,7 +141,6 @@ class ProskommaVerseProvider extends SearchInterface.VerseProvider {
             };
         });
     }
-
     private collectVerseTexts(tokens: GQLBlockToken[]): { [verse: string]: string } {
         return tokens.reduce((verses, token) => {
             const ref = chapterVerseFromScopes(token.scopes);
@@ -191,33 +149,14 @@ class ProskommaVerseProvider extends SearchInterface.VerseProvider {
             return verses;
         }, {});
     }
-
     async queryForBooks(): Promise<GQLBooks> {
-        return (await this.pk.gqlQuery(`
-            {
-                docSet(id: "${this.docSet}") {
-                    documents(${this.searchParams} allChars: true sortedBy: "paratext") {
-                        id
-                        idParts {
-                            type
-                        }
-                    }
-                }
-            }
-        `)) as GQLBooks;
+        return (await this.pk.gqlQuery(
+            ` { docSet(id: "${this.docSet}") { documents(${this.searchParams} allChars: true sortedBy: "paratext") { id idParts { type } } } } `
+        )) as GQLBooks;
     }
-
     async queryForBlocks(bookId: string): Promise<GQLBlocks> {
-        return (await this.pk.gqlQuery(`
-            {
-                document(id: "${bookId}") {
-                    bookCode: header(id: "bookCode")
-                    mainSequence {
-                        blocks(${this.searchParams} allChars: true) {
-                            tokens(includeContext: true) {
-                                scopes(startsWith: ["chapter/" "verses/"])
-                                payload
-                            }
+        return (await this.pk
+            .gqlQuery(` { document(id: "${bookId}") { bookCode: header(id: "bookCode") mainSequence { blocks(${this.searchParams} allChars: true) { tokens(includeContext: true) { scopes(startsWith: ["chapter/" "verses/"]) payload }
                         }
                     }
                 }
@@ -226,65 +165,4 @@ class ProskommaVerseProvider extends SearchInterface.VerseProvider {
     }
 }
 
-interface SearchConfig {
-    ignore?: string;
-    substitute?: { [char: string]: string };
-}
-
-function getConfig(): SearchConfig {
-    return parseConfig(config.mainFeatures['search-accents-to-remove']);
-}
-
-function parseConfig(accentsToRemove: string): SearchConfig {
-    const ignore = parseIgnored(accentsToRemove);
-    const substitute = parseSubstitutes(accentsToRemove);
-    return { ignore, substitute };
-}
-
-function parseSubstitutes(accentsToRemove: string) {
-    const sub = {};
-    for (const match of accentsToRemove.matchAll(/(\S)>(\S)/g)) {
-        extendStringProperty(sub, match[1], match[2]);
-        extendStringProperty(sub, match[2], match[1]);
-    }
-    return sub;
-}
-
-function parseIgnored(accentsToRemove: string) {
-    let ignore = '';
-    for (const c of accentsToRemove.matchAll(/\\u(03\d\d)/g)) {
-        const codePoint = parseInt(c[1], 16);
-        const char = String.fromCodePoint(codePoint);
-        ignore += char;
-    }
-    return ignore;
-}
-
-export const ProksommaSearchInterface = { ProskommaVerseProvider, parseConfig };
-
-/**
- * Implements search queries using Proskomma
- */
-export class SearchQuery extends SearchQueryBase {
-    constructor(
-        searchPhrase: string,
-        pk: SABProskomma,
-        docSet: string,
-        collection: string,
-        options: SearchOptions
-    ) {
-        const configOptions = getConfig();
-        const verseProvider = new ProskommaVerseProvider({
-            pk,
-            searchPhrase,
-            wholeWords: options.wholeWords,
-            ignore: configOptions.ignore,
-            substitute: configOptions.substitute,
-            docSet,
-            collection
-        });
-        options.substitute = configOptions.substitute;
-        options.ignore = configOptions.ignore;
-        super(searchPhrase, verseProvider, options);
-    }
-}
+export const ProksommaSearchInterface = { ProskommaVerseProvider };
